+++ src/argo_interp/cycle/adapter/PchipAdapter.py
@@ -14,13 +14,8 @@ class PchipAdapter(BaseAdapter):
     @classmethod
     def fit(cls, pressure_data: NDArray[np.float64],
             measure_data: NDArray[np.float64],
-            model_kwargs: dict[str, Any] = None,
-            ) -> Self:
-        if model_kwargs is None:
-            model_kwargs = {'extrapolate': False}
-        else:
-            model_kwargs = {'extrapolate': False, **model_kwargs}
-
+            model_kwargs: dict[str, Any]) -> Self:
+        model_kwargs = {'extrapolate': False, **model_kwargs}
         model = PchipInterpolator(x=pressure_data, y=measure_data, **model_kwargs)
         return cls(model=model)